+++ database/migrations/2025_09_28_220517_add_user_id_to_source_controls_table.php
@@ -14,7 +14,10 @@ return new class extends Migration
      */
     public function up(): void
     {
-        DB::statement('DROP INDEX IF EXISTS plugins_is_enabled_priority_index');
+        if (DB::getDriverName() === 'sqlite') {
+            DB::statement('DROP INDEX IF EXISTS plugins_is_enabled_priority_index');
+        }
+
         Schema::table('source_controls', function (Blueprint $table) {
             $table->unsignedBigInteger('user_id')->after('id')->nullable();
         });
